PBL_6조_final/web_frontend/requirements.txt
PyPDF2
pillow
fpdf
Flask
openai
